requirements.txt
opencv-python
pycryptodome
numpy
